Refactor: Implement Profile Fragment, Edit Budget, Logout, and Export Transactions
This commit introduces several enhancements:

- **Profile Management:**
    - Replaced `LainnyaFragment` and `ProfileActivity` with a new `ProfileFragment`.
    - `ProfileFragment` now displays user name and email from `SessionManager`.
    - Implemented logout functionality in `ProfileFragment` which:
        - Calls a new `logout` endpoint in `AuthApiService` and `AuthRepository`.
        - Clears the session using `SessionManager`.
        - Navigates to `SignInActivity`.
    - Added `ic_person.xml` and `ic_logout.xml` drawables for the profile screen.
    - Updated `bottom_nav_menu.xml` to replace "Lainnya" with "Profile".

- **Budget Management:**
    - Added functionality to edit existing budgets.
        - `ManageBudgetFragment` now navigates to `AddBudgetFragment` in edit mode, passing the budget ID.
        - `AddBudgetFragment` can now load budget data for editing, disable category selection in edit mode, and call an update endpoint.
        - `ManageBudgetAdapter` now takes an `onEditClick` listener.
        - `BudgetItem` data class now includes an `id` field and `amount` is `Double`.
    - Adjusted UI elements in `item_budget.xml` (text sizes) and `fragment_add_budget.xml` (title alignment).

- **Transaction Export:**
    - Added functionality to export transaction history.
        - `TabTutupBukuFragment` now has an "Export Laporan" button that calls a new `exportTransactions` endpoint.
        - Added `exportTransactions` function to `TransactionApiService` and `TransactionRepository`.

- **API and Data Model Updates:**
    - Changed `percentage` type from `Int?` to `Double?` in `GetStatAllResponse`, `DetailsItemStat`, and `GetMonthlyBudgetResponse`.
    - Added `logout` endpoint to `AuthApiService`.
    - Added `getRefrestToken` method to `AuthRepository`.
    - Changed `BASE_URL` in `NetworkModule` and updated IP in `network_security_config.xml`.

- **Navigation & Session:**
    - `NavigationActivity` now checks login status using `SessionManager` on `onCreate` and redirects to `SignInActivity` if not logged in.

- **Minor Code Refinements:**
    - Updated `TabTutupBukuFragment` to use `CategoryIconMapper` and pass budget ID when creating `BudgetItem`.
    - Ensured `TransactionRepository` uses the correct `getAuthToken()` method.
    - General improvements to toast messages and loading indicators.

+++ app/src/main/java/com/fintrack/app/data/response/DetailsItemStat.kt
@@ -10,5 +10,5 @@ data class DetailsItemStat(
     val totalAmount: Int? = null,
 
     @field:SerializedName("percentage")
-    val percentage: Int? = null
+    val percentage: Double? = null
 )
